Renaming form components
